modify v alias in .bashrc; Firefox dialog doesn't warp; many Xemacs changes
v alias now views directories too.

At work, the timesheet webpage would always create a dialog when it
automatically logged me out.  Fvwm would warp to the dialog and focus
it, which was very distracting.  This is now fixed with a splattering
of Styles; some could be eliminated.

In Xemacs:
   color-theme of choice is now moved to my ~/.xemacs and modified so
   it doesn't turn paren-mode on (paren-mode is broken).

   C-<tab> is not bound to anything

   added redo package to get simple undo/redo